Even moreeeeeeeee commitsssss

+++ pow.c
@@ -2,18 +2,18 @@
 #include <math.h>
 int main()
 {
-    int x,r;
-    int n;
+    double x,r;
+    double n;
 
     printf("Enter the x number: ");
     scanf("%lf", &x);
 
     printf("Enter the n number: ");
-    scanf("%d",&n);
+    scanf("%lf",&n);
 
-    r= pow(n,x);
+    r = pow(n,x);
 
-    printf("%.1lf^%.1d = %.2lf", x, n, r);
+    printf("%.1lf^%.1f = %.2lf", x, n, r);
 
     return 0;
 }
